Refactor type casting and improve null checks for better stability
- Updated type casting in `PhysicalizeStaticObject` and `DrawEntity` methods to use `intptr_t` for safer pointer arithmetic.
- Enhanced null checks in `PreloadInstanceResources` and `OnActivate` methods to streamline conditional logic and improve readability.
- Refined matrix manipulation in `DrawEntity` to utilize `SetTranslationMat` for clarity and maintainability.
- Improved debug rendering in `LoadVisAreaBoxFromXML` and `PortalsDrawDebug` by ensuring consistent pointer checks and using local color variables for better code organization.
- Addressed assert handling in `PreCacheArea` to ensure proper behavior in release mode.

+++ Cry3DEngine/3dEngine.cpp
@@ -860,7 +860,7 @@ bool C3DEngine::PhysicalizeStaticObject(void *pForeignData,int iForeignData,int 
 	if (!m_pTerrain || !m_pObjManager)
 		return false;
 
-	int ix=(int)pForeignData&0xFF, iy=(int)pForeignData>>8&0xFF, iobj=(int)pForeignData>>16&0xFFFF;
+	int ix=(int)(intptr_t)pForeignData&0xFF, iy=(int)(intptr_t)pForeignData>>8&0xFF, iobj=(int)(intptr_t)pForeignData>>16&0xFFFF;
 	return 0;//m_pObjManager->PhysicalizeStatObjInst( &m_pTerrain->m_arrSecInfoTable[ix][iy]->m_lstStatObjects[iobj], true );
 }
